+++ include/fa/svgs/brands/index.php
@@ -1,7 +1,0 @@
-<?php
-/*
- * SPDX-FileCopyrightText: 2004-2026 The Cacti Group
- * SPDX-License-Identifier: GPL-2.0-or-later
- */
-
-header("Location:../index.php");