+++ public/index.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * SPDX-FileCopyrightText: 2026 The Kadupul project and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+use Symfony\Component\HttpFoundation\Request;
+
+$kernel = require dirname(__DIR__) . '/config/bootstrap.php';
+$request = Request::createFromGlobals();
+$response = $kernel->handle($request);
+$response->send();
+$kernel->terminate($request, $response);