+++ tests/security/build_architectural_helper_report.sh
@@ -13,6 +13,8 @@ MODE="${1:---summary}"
 
 EXCLUDE=(
 	--glob '!include/vendor/**'
+	--glob '!node_modules/**'
+	--glob '!var/**'
 	--glob '!locales/**'
 	--glob '!tests/**'
 	--glob '!**/*.min.js'